ASTPrinter: Refactor the definition of SynthesizedExtensionAnalyzer to IDETypeChecking.cpp. NFC (#10975)
